+++ gempak/source/programs/dc/dcffa/dcffa.c
@@ -13,7 +13,7 @@
 
 void fa_dcod ( char *curtim, char *gemfil, char *stntbl, char *stntb2, 
 	       char *prmfil, int *iadstn, int *maxtim, int *nhours, 
-	       int *iret, int, int, int, int, int );
+	       int *iret, size_t, size_t, size_t, size_t, size_t );
 
 int main ( int argc, char *argv[] )
 /************************************************************************
@@ -28,7 +28,8 @@ int main ( int argc, char *argv[] )
  **									*
  * Log:									*
  * H. Zeng/SAIC		07/05						*
- * L. Hinson/AWC        06/08   Added circflg parameter to dc_gopt      *	
+ * L. Hinson/AWC        06/08   Added circflg parameter to dc_gopt      *
+ * SGuan-BHebbard/NCEP  06/22   Changed strlen in protos int->size_t    *	
  ***********************************************************************/
 {
