+++ gempak/source/programs/dc/dcidft/dcidft.c
@@ -13,7 +13,7 @@
 
 void si_dcod ( char *curtim, char *gemfil, char *icetbl, char *prmfil, 
 		  int *iadstn, int *maxtim, int *nhours, int *iret, 
-                  int, int, int, int );
+                  size_t, size_t, size_t, size_t );
 
 int main ( int argc, char *argv[] )
 /************************************************************************
@@ -33,6 +33,7 @@ int main ( int argc, char *argv[] )
  * L. Hinson/AWC        06/08   Added circflg parameter to dc_gopt      *
  * S. Jacobs/NCEP	 5/13	Increased the number of stations from	*
  * 				500 to 1200				*
+ * S. Guan/NCEP		 6/22	Fixed the length of Strings		*
  ***********************************************************************/
 {
