+++ gempak/source/programs/gd/gdprof/gdpduv.c
@@ -6,25 +6,25 @@ void gdpduvf ( int *iflno, char *gdatim, char *gvcord, char *gvect,
                float *ystop, float *gridu, float *gridv, float *rgx, 
                float *rgy, float *rlat, float *rlon,
                int *npts, float *u, float *v, float *y, char *parmu, 
-               char *parmv, int *iret, int gdatimsize, int gvcordsize, 
-               int gvectsize, int gpointsize, int timesize, int parmusize, 
-               int parmvsize );
+               char *parmv, int *iret, size_t gdatimsize, size_t gvcordsize, 
+               size_t gvectsize, size_t gpointsize, size_t timesize, size_t parmusize, 
+               size_t parmvsize );
 
 void gdpduv  ( int *iflno, char *gdatim, char *gvcord, char *gvect, 
                char *gpoint, char *time, int *ivcord, float *ystrt, 
                float *ystop, float *rgx, float *rgy, float *rlat, float *rlon,
                int *npts, float *u, float *v, float *y, char *parmu, 
-               char *parmv, int *iret, int gdatimsize, int gvcordsize, 
-               int gvectsize, int gpointsize, int timesize, int parmusize, 
-               int parmvsize );
+               char *parmv, int *iret, size_t gdatimsize, size_t gvcordsize, 
+               size_t gvectsize, size_t gpointsize, size_t timesize, size_t parmusize, 
+               size_t parmvsize );
 
 void gdpduv  ( int *iflno, char *gdatim, char *gvcord, char *gvect, 
                char *gpoint, char *time, int *ivcord, float *ystrt, 
                float *ystop, float *rgx, float *rgy, float *rlat, float *rlon,
                int *npts, float *u, float *v, float *y, char *parmu, 
-               char *parmv, int *iret, int gdatimsize, int gvcordsize, 
-               int gvectsize, int gpointsize, int timesize, int parmusize, 
-               int parmvsize )
+               char *parmv, int *iret, size_t gdatimsize, size_t gvcordsize, 
+               size_t gvectsize, size_t gpointsize, size_t timesize, size_t parmusize, 
+               size_t parmvsize )
 /************************************************************************
  * gdpduv								*
  *									*
@@ -69,6 +69,7 @@ void gdpduv  ( int *iflno, char *gdatim, char *gvcord, char *gvect,
  * S. Gilbert/NCEP	08/07   					*
  * S. Jacobs/NCEP	02/08	Set grid size as LLMXGD			*
  * T. Piper/SAIC	03/08	Replaced cmm functions with Macros	*
+ * B. Hebbard/NCEP	08/21	Changed strlen in protos int->size_t    *
  ***********************************************************************/
 {
     int   kx, ky, kxy, ier;
