+++ gempak/source/programs/gd/gdthgt/gdtxrd.c
@@ -4,15 +4,15 @@
 void gdtxrdf ( char *time, char *gvcord, char *gfunc, int *levt, int *levb, 
               int *lavflg, float *rgx, float *rgy, int *inttyp, 
               float *grid, float *yy, 
-              int *iret, int timesize, int gvcordsize, int gfuncsize );
+              int *iret, size_t timesize, size_t gvcordsize, size_t gfuncsize );
 
 void gdtxrd ( char *time, char *gvcord, char *gfunc, int *levt, int *levb, 
               int *lavflg, float *rgx, float *rgy, int *inttyp, float *yy, 
-              int *iret, int timesize, int gvcordsize, int gfuncsize );
+              int *iret, size_t timesize, size_t gvcordsize, size_t gfuncsize );
 
 void gdtxrd ( char *time, char *gvcord, char *gfunc, int *levt, int *levb, 
               int *lavflg, float *rgx, float *rgy, int *inttyp, float *yy, 
-              int *iret, int timesize, int gvcordsize, int gfuncsize )
+              int *iret, size_t timesize, size_t gvcordsize, size_t gfuncsize )
 /************************************************************************
  * gdtxrd								*
  *									*
@@ -43,6 +43,7 @@ void gdtxrd ( char *time, char *gvcord, char *gfunc, int *levt, int *levb,
  * Log:									*
  * S.Gilbert/NCEP	 8/07	                       			*
  * T. Piper/SAIC	03/08	Replaced cmm functions with Macros	*
+ * B. Hebbard/NCEP	08/21	Changed strlen in protos int->size_t    *
  ***********************************************************************/
 {
     int   kx, ky, kxy, ier;
